+++ src/app/(menu)/elite/page.tsx
@@ -2,7 +2,7 @@ import { service } from '@/service/pickService';
 import styles from './page.module.css';
 import EliteList from '@/components/routes/elite/EliteList';
 import DragAndDropProvider from '@/context/DragAndDrop/DragAndDropProvider';
-import Notice from '@/components/common/Notice/Notice';
+import Notice from '@/components/common/Notice';
 import { Metadata } from 'next';
 
 export const metadata: Metadata = {
